+++ lib/buffer/reflow_strategy.dart
@@ -1,5 +1,5 @@
 import 'package:xterm/buffer/buffer.dart';
-import 'package:xterm/buffer/buffer_line.dart';
+import 'package:xterm/buffer/line/line.dart';
 import 'package:xterm/util/circular_list.dart';
 
 abstract class ReflowStrategy {