+++ lib/buffer/reflow_strategy_narrower.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:xterm/buffer/buffer.dart';
-import 'package:xterm/buffer/buffer_line.dart';
+import 'package:xterm/buffer/line/line.dart';
 import 'package:xterm/buffer/reflow_strategy.dart';
 
 class ReflowStrategyNarrower extends ReflowStrategy {