+++ lib/xterm.dart
@@ -1,7 +1,7 @@
 library xterm;
 
 export 'terminal/terminal.dart';
-export 'terminal/terminal_isolate.dart';
 export 'terminal/terminal_backend.dart';
 export 'terminal/terminal_ui_interaction.dart';
 export 'terminal/platform.dart';
+export 'theme/terminal_style.dart';